Revert "Implemented Scriptable Objects for Food-Types and Integrated Cost Management System"
This reverts commit 26813bbc5fce81c4173a2896f445c5b9a74b1cf4.

+++ Assets/Scripts/Game Economy/FoodProperties.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "FoodProperties", menuName = "Game Economy/Food Properties")]
 public class FoodProperties : ScriptableObject
 {
-    public int foodCost;
-    public int damage;
+    public float foodCost;
+    public float damage;
     public float staminaTime;
     public float destructionTime;
 }
